requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0
rich>=13.7.0
